Changes for image output and track reports. Added new test case.

+++ java/TikaImageDetection/src/main/java/org/mitre/mpf/detection/tika/EmbeddedContentExtractor.java
@@ -28,10 +28,9 @@ package org.mitre.mpf.detection.tika;
 import org.apache.tika.extractor.EmbeddedDocumentExtractor;
 import org.apache.tika.metadata.Metadata;
 
+
+import org.slf4j.Logger;
 import org.xml.sax.ContentHandler;
-import org.xml.sax.SAXException;
-import org.xml.sax.SAXParseException;
-import org.xml.sax.Attributes;
 
 import java.io.File;
 import java.io.InputStream;
@@ -42,40 +41,55 @@ import java.nio.file.Path;
 import java.lang.StringBuilder;
 
 import java.util.ArrayList;
-import java.util.Set;
-import java.util.HashSet;
 import java.util.HashMap;
 
 
 
 public class EmbeddedContentExtractor implements EmbeddedDocumentExtractor {
     private String path;
-    private HashMap<String, String> imagesFound;
-    private boolean separatePages;
-    private int id,pagenum;
+    private HashMap<String, String> imagesFound, imagesIndex;
+    private ArrayList<String> commonImages;
+    private boolean separatePages, repeatImages;
+    private int id, pagenum;
     private ArrayList<StringBuilder> imageMap;
     private StringBuilder current;
-    private Path outputPageDir;
-    public EmbeddedContentExtractor(String savePath, boolean separate){
+    private Path outputDir;
+    private Path commonImgDir;
+    private Logger LOG;
+    public EmbeddedContentExtractor(String savePath, boolean separate, boolean repeat) {
         path = savePath;
-        pagenum = 0;
-        imagesFound = new HashMap<String,String>();
+        pagenum = -1;
+        imagesFound = new HashMap<String, String>();
+        imagesIndex = new HashMap<String, String>();
+        commonImages = new ArrayList<String>();
         separatePages = separate;
+        repeatImages = repeat;
         id = 0;
         imageMap = new ArrayList<StringBuilder>();
         current = new StringBuilder();
-        imageMap.add(current);
-        if(separatePages)
-            outputPageDir = (new File(path + "/PAGE_" + String.valueOf(pagenum-1))).toPath();
 
-        Path outputDir = (new File(path)).toPath();
-        outputPageDir = outputDir;
+        outputDir = (new File(path + "/tika-extracted")).toPath();
+        commonImgDir =  (new File(path + "/tika-extracted/common")).toPath();
+
+    }
+    public String init(Logger log) {
+        LOG = log;
+        String errMsg = "";
         try {
-            if (!Files.exists(outputDir))
+            if (!Files.exists(outputDir)) {
                 Files.createDirectories(outputDir);
+            }
+            if (repeatImages && !Files.exists(commonImgDir)) {
+                Files.createDirectories(commonImgDir);
+            }
         } catch (IOException e) {
             e.printStackTrace();
+            errMsg = String.format("Unable to create the following directories: \n%s\n%s",
+                    outputDir.toAbsolutePath(), commonImgDir.toAbsolutePath());
+            LOG.error(errMsg, e);
         }
+        return errMsg;
+
     }
 
     public ArrayList<StringBuilder> getImageList() {
@@ -87,38 +101,68 @@ public class EmbeddedContentExtractor implements EmbeddedDocumentExtractor {
     }
 
     public void parseEmbedded(InputStream stream, ContentHandler imHandler, Metadata metadata, boolean outputHtml)
-            throws SAXException, IOException {
+            throws IOException {
         //Create a new page
         int nextpage = Integer.parseInt(imHandler.toString()) - 1;
         while (pagenum < nextpage) {
-            pagenum ++;
+            pagenum++;
             current = new StringBuilder();
             imageMap.add(current);
             if (separatePages) {
-                outputPageDir = (new File(path + "/PAGE_" + String.valueOf(pagenum))).toPath();
+                outputDir = (new File(path + "/page-" + String.valueOf(pagenum))).toPath();
             }
-            Files.createDirectories(outputPageDir);
+            Files.createDirectories(outputDir);
         }
 
 
         String cosID = metadata.get(Metadata.EMBEDDED_RELATIONSHIP_ID);
-        if (imagesFound.containsKey(cosID)) {
-            //Skip saving the file but append the file information.
+        String filename = "image" + String.valueOf(id) + "." + metadata.get(Metadata.CONTENT_TYPE);
+        if (imagesFound.containsKey(cosID) ) {
+            if (repeatImages && !commonImages.contains(cosID)) {
+                // For images already encountered, save into a common images folder.
+                // Save each image only once in the common images folder.
+                commonImages.add(cosID);
+                String imageFile = imagesIndex.get(cosID);
+                File outputFPath = new File(commonImgDir.toString() + "/" + imageFile);
+                Path outputPath = outputFPath.toPath();
+
+
+
+                if (Files.exists(outputPath)) {
+                    // Warn users that folder already contains images being overwritten
+                    // This only happens if the same jobID was used on separate runs.
+                    LOG.warn("File {} already exists and will be overwritten", outputPath.toAbsolutePath());
+                    Files.deleteIfExists(outputPath);
+                }
+                Files.copy(stream, outputPath);
+
+                filename = outputFPath.getAbsolutePath();
+                imagesFound.put(cosID, filename);
+            } else {
+                filename = imagesFound.get(cosID);
+            }
         } else {
-            imagesFound.put(cosID,"image" + String.valueOf(id) + "." + metadata.get(Metadata.CONTENT_TYPE));
-            File outputFPath = new File(outputPageDir.toString() + "/image" + String.valueOf(id) + "." + metadata.get(Metadata.CONTENT_TYPE));
+            imagesIndex.put(cosID, filename);
+            File outputFPath = new File(outputDir.toString() + "/" + filename);
             Path outputPath = outputFPath.toPath();
-            Files.deleteIfExists(outputPath);
-            boolean check = Files.exists(outputPath);
+
+            if (Files.exists(outputPath)) {
+                // Warn users that folder already contains images being overwritten
+                // This only happens if the same jobID was used on separate runs.
+                LOG.warn("File {} already exists and will be overwritten", outputPath.toAbsolutePath());
+                Files.deleteIfExists(outputPath);
+            }
             Files.copy(stream, outputPath);
+            filename = outputFPath.getAbsolutePath();
+            imagesFound.put(cosID, filename);
             id++;
         }
 
-        String filename = imagesFound.get(cosID);
         if (current.length() == 0) {
             current.append(filename);
         } else {
-            current.append(", " + filename);
+            current.append("; " + filename);
         }
+
     }
 }
